iac/aws/iac/pyproject.toml
[tool.poetry]
name = "raa.datalake"
version = "0.0.1"
description = "AWS CDK infrastructure project for the Retail Analytics datalake stack"
authors = ["AI Alliance Devops Team <devops@thealliance.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.150.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.2"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"